+++ src/ServerConfig.cc
@@ -194,8 +194,13 @@ class ignition::gazebo::ServerConfigPrivate
             updateRate(_cfg->updateRate),
             useLevels(_cfg->useLevels),
             useDistributed(_cfg->useDistributed),
+            useLogRecord(_cfg->useLogRecord),
+            logRecordPath(_cfg->logRecordPath),
+            logPlaybackPath(_cfg->logPlaybackPath),
             resourceCache(_cfg->resourceCache),
-            plugins(_cfg->plugins) { }
+            plugins(_cfg->plugins),
+            networkRole(_cfg->networkRole),
+            networkSecondaries(_cfg->networkSecondaries) { }
 
   // \brief The SDF file that the server should load
   public: std::string sdfFile = "";
@@ -212,12 +217,27 @@ class ignition::gazebo::ServerConfigPrivate
   /// \brief Use the distributed simulation system
   public: bool useDistributed{false};
 
+  /// \brief Use the logging system to record states
+  public: bool useLogRecord{false};
+
+  /// \brief Path to place recorded states
+  public: std::string logRecordPath = "";
+
+  /// \brief Path to recorded states to play back using logging system
+  public: std::string logPlaybackPath = "";
+
   /// \brief Path to where simulation resources, such as models downloaded
   /// from fuel.ignitionrobotics.org, should be stored.
   public: std::string resourceCache = "";
 
   /// \brief List of plugins to load.
   public: std::list<ServerConfig::PluginInfo> plugins;
+
+  /// \brief The network role.
+  public: std::string networkRole = "";
+
+  /// \brief The number of network secondaries.
+  public: unsigned int networkSecondaries = 0;
 };
 
 //////////////////////////////////////////////////
@@ -303,15 +323,79 @@ void ServerConfig::SetUseLevels(const bool _levels)
 }
 
 /////////////////////////////////////////////////
+void ServerConfig::SetNetworkSecondaries(unsigned int _secondaries)
+{
+  this->dataPtr->networkSecondaries = _secondaries;
+}
+
+/////////////////////////////////////////////////
+unsigned int ServerConfig::NetworkSecondaries() const
+{
+  return this->dataPtr->networkSecondaries;
+}
+
+/////////////////////////////////////////////////
+void ServerConfig::SetNetworkRole(const std::string &_role)
+{
+  this->dataPtr->networkRole = _role;
+}
+
+/////////////////////////////////////////////////
+std::string ServerConfig::NetworkRole() const
+{
+  return this->dataPtr->networkRole;
+}
+
+/////////////////////////////////////////////////
 bool ServerConfig::UseDistributedSimulation() const
 {
-  return this->dataPtr->useDistributed;
+  // We just check that network role is not empty.
+  // src/network/NetworkConfig.cc checks if this value is valid.
+  // \todo(nkoenig) Deprecated "SetUseDistributedSimulation" in ign-gazebo2
+  // and remove the "|| this->dataPtr->useDistributed" in ign-gazebo3.
+  return !this->dataPtr->networkRole.empty() || this->dataPtr->useDistributed;
 }
 
 /////////////////////////////////////////////////
 void ServerConfig::SetUseDistributedSimulation(const bool _distributed)
 {
   this->dataPtr->useDistributed = _distributed;
+}
+
+/////////////////////////////////////////////////
+bool ServerConfig::UseLogRecord() const
+{
+  return this->dataPtr->useLogRecord;
+}
+
+/////////////////////////////////////////////////
+void ServerConfig::SetUseLogRecord(const bool _record)
+{
+  this->dataPtr->useLogRecord = _record;
+}
+
+/////////////////////////////////////////////////
+const std::string ServerConfig::LogRecordPath() const
+{
+  return this->dataPtr->logRecordPath;
+}
+
+/////////////////////////////////////////////////
+void ServerConfig::SetLogRecordPath(const std::string &_recordPath)
+{
+  this->dataPtr->logRecordPath = _recordPath;
+}
+
+/////////////////////////////////////////////////
+const std::string ServerConfig::LogPlaybackPath() const
+{
+  return this->dataPtr->logPlaybackPath;
+}
+
+/////////////////////////////////////////////////
+void ServerConfig::SetLogPlaybackPath(const std::string &_playbackPath)
+{
+  this->dataPtr->logPlaybackPath = _playbackPath;
 }
 
 /////////////////////////////////////////////////