+++ src/SyncManager.cc
@@ -15,24 +15,20 @@
  *
  */
 
-#include "ignition/common/Profiler.hh"
-#include "ignition/gazebo/Events.hh"
-#include "ignition/gazebo/EntityComponentManager.hh"
+#include <ignition/common/Profiler.hh>
+#include <ignition/gazebo/Events.hh>
 
+#include "ignition/gazebo/EntityComponentManager.hh"
 #include "ignition/gazebo/components/Name.hh"
 #include "ignition/gazebo/components/ParentEntity.hh"
 #include "ignition/gazebo/components/Performer.hh"
 #include "ignition/gazebo/components/PerformerAffinity.hh"
-#include "ignition/gazebo/components/Pose.hh"
-#include "ignition/gazebo/components/Static.hh"
-#include "ignition/gazebo/components/World.hh"
 
 #include "SyncManager.hh"
 #include "SimulationRunner.hh"
 
 #include "network/NetworkManagerPrimary.hh"
 #include "network/NetworkManagerSecondary.hh"
-#include "network/components/PerformerActive.hh"
 
 #include "msgs/performer_affinity.pb.h"
 
@@ -60,12 +56,13 @@ SyncManager::SyncManager(
 
   if (this->role != NetworkRole::None)
   {
-    this->posePub = this->node.Advertise<ignition::msgs::Pose_V>("pose_update");
+    this->statePub = this->node.Advertise<msgs::SerializedState>(
+        "state_update");
   }
 
   if (this->role == NetworkRole::SimulationPrimary)
   {
-    this->node.Subscribe("pose_update", &SyncManager::OnPose, this);
+    this->node.Subscribe("state_update", &SyncManager::OnState, this);
   }
 }
 
@@ -79,9 +76,8 @@ void SyncManager::DistributePerformers()
 
     auto &secondaries = mgr->Secondaries();
     auto secondaryIt = secondaries.begin();
-    auto &ecm = this->runner->entityCompMgr;
 
-    msgs::PerformerAffinities msg;
+    private_msgs::PerformerAffinities msg;
 
     this->runner->entityCompMgr.Each<components::Performer,
                                      components::ParentEntity>(
@@ -98,12 +94,6 @@ void SyncManager::DistributePerformers()
         affinityMsg->mutable_entity()->set_id(_entity);
         affinityMsg->set_secondary_prefix(secondaryIt->second->prefix);
 
-        auto isStatic = ecm.Component<components::Static>(pid);
-        *isStatic = components::Static(false);
-
-        auto isActive = ecm.Component<components::PerformerActive>(_entity);
-        *isActive = components::PerformerActive(true);
-
         this->runner->entityCompMgr.CreateComponent(_entity,
         components::PerformerAffinity(secondaryIt->second->prefix));
 
@@ -118,27 +108,46 @@ void SyncManager::DistributePerformers()
 
     for (auto &secondary : secondaries)
     {
-      bool result;
-      msgs::PerformerAffinities resp;
-
       std::string topic {secondary.second->prefix + "/affinity"};
       unsigned int timeout = 5000;
 
-      bool executed = this->node.Request(topic, msg, timeout, resp, result);
-      if (executed)
+      std::vector<transport::ServicePublisher> publishers;
+      const std::size_t tries = 30;
+      for (size_t i = 0; i < tries; ++i)
       {
-        if (!result)
-        {
+        this->node.ServiceInfo(topic, publishers);
+        if (!publishers.empty())
+          break;
+        std::this_thread::sleep_for(std::chrono::milliseconds(timeout/tries));
+      }
+
+      if (publishers.empty())
+      {
           ignwarn << "Failed to set performer affinities for " <<
-            secondary.second->prefix << " (service call failed)" <<
+            secondary.second->prefix << " (service not available)" <<
             std::endl;
-        }
       }
       else
       {
-        ignwarn << "Failed to set performer affinities for " <<
-          secondary.second->prefix << " (service call timed out)" <<
-          std::endl;
+        bool result;
+        private_msgs::PerformerAffinities resp;
+
+        bool executed = this->node.Request(topic, msg, timeout, resp, result);
+        if (executed)
+        {
+          if (!result)
+          {
+            ignwarn << "Failed to set performer affinities for " <<
+              secondary.second->prefix << " (service call failed)" <<
+              std::endl;
+          }
+        }
+        else
+        {
+          ignwarn << "Failed to set performer affinities for " <<
+            secondary.second->prefix << " (service call timed out)" <<
+            std::endl;
+        }
       }
     }
   }
@@ -150,37 +159,33 @@ void SyncManager::DistributePerformers()
     std::string topic {mgr->Namespace() + "/affinity"};
     bool received = false;
 
-    std::function<bool(const msgs::PerformerAffinities &,
-                       msgs::PerformerAffinities &)> fcn =
-      [&received, &mgr, this, &ecm](const msgs::PerformerAffinities &_req,
-        msgs::PerformerAffinities &/*_resp*/) -> bool
+    std::function<bool(const private_msgs::PerformerAffinities &,
+                       private_msgs::PerformerAffinities &)> fcn =
+      [&received, &mgr, this, &ecm](
+        const private_msgs::PerformerAffinities &_req,
+        private_msgs::PerformerAffinities &/*_resp*/) -> bool
       {
         for (int ii = 0; ii < _req.affinity_size(); ++ii)
         {
           const auto &affinityMsg = _req.affinity(ii);
           const auto &entityId = affinityMsg.entity().id();
 
-          auto pid = ecm.Component<components::ParentEntity>(entityId);
+          auto pid = ecm.Component<components::ParentEntity>(entityId)->Data();
 
           ecm.CreateComponent(entityId,
             components::PerformerAffinity(affinityMsg.secondary_prefix()));
 
-          auto isStatic = ecm.Component<components::Static>(pid->Data());
-          auto isActive = ecm.Component<components::PerformerActive>(entityId);
-
           if (affinityMsg.secondary_prefix() == mgr->Namespace())
           {
             this->performers.push_back(entityId);
-            *isStatic = components::Static(false);
-            *isActive = components::PerformerActive(true);
             igndbg << "Secondary [" << mgr->Namespace()
                    << "] assigned affinity to performer [" << entityId << "]."
                    << std::endl;
           }
+          // Remove performers not handled by this secondary
           else
           {
-            *isStatic = components::Static(true);
-            *isActive = components::PerformerActive(false);
+            ecm.RequestRemoveEntity(pid);
           }
         }
         received = true;
@@ -191,19 +196,20 @@ void SyncManager::DistributePerformers()
 
     igndbg << "Secondary [" << mgr->Namespace()
            << "] waiting for affinity assignment." << std::endl;
-    while (!received &&
-           !this->runner->stopReceived)
+    while (!received && !this->runner->stopReceived)
     {
       std::this_thread::sleep_for(std::chrono::milliseconds(1));
     }
+    igndbg << "Secondary [" << mgr->Namespace()
+           << "] received affinity assignment." << std::endl;
   }
 }
 
 /////////////////////////////////////////////////
-void SyncManager::OnPose(const ignition::msgs::Pose_V & _msg)
+void SyncManager::OnState(const msgs::SerializedState & _msg)
 {
-  std::lock_guard<std::mutex> lock(this->poseMutex);
-  this->poseMsgs.push_back(_msg);
+  std::lock_guard<std::mutex> lock(this->msgMutex);
+  this->stateMsgs.push_back(_msg);
 }
 
 /////////////////////////////////////////////////
@@ -211,38 +217,28 @@ bool SyncManager::Sync()
 {
   IGN_PROFILE("SyncManager::Sync");
 
-  // TODO(mjcarroll) this is where more advanced serialization/sync will go.
-  auto& ecm = this->runner->entityCompMgr;
+  auto &ecm = this->runner->entityCompMgr;
 
   if (this->role == NetworkRole::SimulationSecondary)
   {
-    ignition::msgs::Pose_V msg;
-
-    for (const auto &entity : this->performers)
+    // Get all the performer's models
+    std::unordered_set<Entity> models;
+    for (const auto &perf : this->performers)
     {
-      auto pid = ecm.Component<components::ParentEntity>(entity);
-      auto pose = ecm.Component<components::Pose>(pid->Data());
-      auto poseMsg = msg.add_pose();
-      ignition::msgs::Set(poseMsg, pose->Data());
-      poseMsg->set_id(entity);
+      models.insert(ecm.Component<components::ParentEntity>(perf)->Data());
     }
-    this->posePub.Publish(msg);
+
+    auto msg = ecm.State(models);
+    this->statePub.Publish(msg);
   }
   else
   {
-    std::lock_guard<std::mutex> lock(this->poseMutex);
-    for (const auto& msg : this->poseMsgs)
+    std::lock_guard<std::mutex> lock(this->msgMutex);
+    for (const auto &msg : this->stateMsgs)
     {
-      for (int ii = 0; ii < msg.pose_size(); ++ii)
-      {
-        const auto& poseMsg = msg.pose(ii);
-        auto pid = ecm.Component<components::ParentEntity>(poseMsg.id());
-        auto pose = ecm.Component<components::Pose>(pid->Data());
-        auto newPose = ignition::msgs::Convert(poseMsg);
-        *pose = components::Pose(newPose);
-      }
+      ecm.SetState(msg);
     }
-    this->poseMsgs.clear();
+    this->stateMsgs.clear();
   }
   return true;
 }