+++ Changelog.md
@@ -1,3 +1,58 @@
+## Ignition Gazebo 1.x
+
+### Ignition Gazebo 1.X.X
+
+1. Add a basic JointController system
+    * [Pull Request 246](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/246)
+
+1. Added command line options to configure distributed simulation. These
+   will replace the environment variables.
+    * [Pull Request 238](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/238)
+
+
+### Ignition Gazebo 1.1.0 (2019-03-15)
+
+1. Distributed performers running in lockstep
+    * [Pull Request 186](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/186)
+    * [Pull Request 201](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/201)
+    * [Pull Request 209](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/209)
+    * [Pull Request 213](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/213)
+
+1. Fix documentation tagfiles
+    * [Pull Request 214](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/214)
+
+1. Convert gui library into a component
+    * [Pull Request 206](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/206)
+
+1. include <cstdint> wherever special int types like uint64_t are used
+    * [Pull Request 208](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/208)
+
+1. Move network internal
+    * [Pull Request 211](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/211)
+
+1. Logging / playback
+    * [Pull Request 181](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/181)
+
+1. ECM state streaming
+    * [Pull Request 184](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/184)
+
+1. Unversioned system libraries
+    * [Pull Request 222](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/222)
+
+### Ignition Gazebo 1.0.2 (2019-03-12)
+
+1. Use TARGET_SO_NAME to fix finding dartsim plugin
+    * [Pull Request 217](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/217)
+
+### Ignition Gazebo 1.0.1 (2019-03-01)
+
+1. Update gazebo version number in sdf files
+    * [Pull Request 207](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/207)
+
+### Ignition Gazebo 1.0.0 (2019-03-01)
+
+1. Initial release
+
 ## Ignition Gazebo 0.x
 
 ### Ignition Gazebo 0.1.0